fix: Missing const on TwoDA
fix: Move in Requirement when constructing a Modifier
fix: Misspelling in Class
fix: Rules::replace pass Requirement by const ref
fix: sign mismatch in LevelStats

+++ lib/nw/objects/components/LevelStats.cpp
@@ -77,7 +77,7 @@ int LevelStats::level() const noexcept
 int LevelStats::level_by_class(Index id) const noexcept
 {
     for (const auto& cl : entries) {
-        if (*id == cl.id) {
+        if (*id == static_cast<size_t>(cl.id)) {
             return cl.level;
         }
     }
